Add a test scene, with a picocar with a script that just lets us choose a body colour in the editor
This was harder than expected. The key thing was understanding that the scripts would not be active in the editor itself and need @tool to make changes in the editor. With @tool the changes to exported variables go through our getter/setter and appropriately update the material for the car.